# docs/swagger.yaml
basePath: /todos
definitions:
  main.Todo:
    properties:
      complete:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo API
  version: "1.0"
paths:
  /todo:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new todo with the input payload
      parameters:
      - description: Create todo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
      summary: Create a new todo
      tags:
      - Todos
  /todos:
    get:
      description: Get details of all todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Todo'
            type: array
      summary: Get details of all todos
      tags:
      - Todos
  /todos/{id}:
    delete:
      description: Delete specific todo
      parameters:
      - description: Delete todos by Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
      summary: Delete specific todo
      tags:
      - Todos
    get:
      description: Get specific todo
      parameters:
      - description: Get todos by Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
      summary: Get specific todo
      tags:
      - Todos
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Update specific todo
      parameters:
      - description: Update todos by Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Todo'
      summary: Update specific todo
      tags:
      - Todos
swagger: "2.0"